Move from dataformat to datamodel
