--- morphir/dataformat/src/org/finos/morphir/datamodel/Label.scala
@@ -1,3 +1,0 @@
-package org.finos.morphir.datamodel
-
-case class Label(value: String)
