+++ morphir/datamodel/src/org/finos/morphir/datamodel/Attributes.scala
@@ -1,0 +1,10 @@
+package org.finos.morphir.datamodel
+trait Attributes {
+  def isEmpty: Boolean
+}
+
+object Attributes {
+  sealed trait Binding extends Product with Serializable {
+    def keyId: ConceptId
+  }
+}
