+++ morphir/datamodel/src/org/finos/morphir/datamodel/Label.scala
@@ -1,0 +1,2 @@
+package org.finos.morphir.datamodel
+case class Label(value: String)
